parent login done

+++ app/Http/Controllers/HomeworkController.php
@@ -34,6 +34,7 @@ class HomeworkController extends Controller
      */
     public function create()
     {
+        $this->authorize('is-admin');
         return view('add-homework');
     }
 
@@ -45,6 +46,7 @@ class HomeworkController extends Controller
      */
     public function store(StoreHomeworkRequest $request)
     {
+        $this->authorize('is-admin');
         $request->validate([
             'date' => ['required', 'date'],
             'class' => ['required', 'string', 'max:255'],
@@ -81,6 +83,7 @@ class HomeworkController extends Controller
      */
     public function edit(Homework $homework)
     {
+        $this->authorize('is-admin');
         $homework = Homework::find($homework)->first();
         return view('edit-homework',[
             'homework' => $homework
@@ -96,6 +99,7 @@ class HomeworkController extends Controller
      */
     public function update(UpdateHomeworkRequest $request, Homework $homework)
     {
+        $this->authorize('is-admin');
         $request->validate([
             'date' => ['required', 'date'],
             'class' => ['required', 'string', 'max:255'],
@@ -121,6 +125,7 @@ class HomeworkController extends Controller
      */
     public function destroy(Homework $homework)
     {
+        $this->authorize('is-admin');
         $homework->delete();
         return redirect('/homeworks')->with('success','Homework deleted successfully!');
     }
